+++ src/exceptions.rs
@@ -30,7 +30,7 @@ macro_rules! impl_exception_boilerplate {
         impl std::error::Error for $name {
             fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
                 unsafe {
-                    use $crate::{AsPyPointer, PyNativeType};
+                    use $crate::AsPyPointer;
                     let cause: &$crate::exceptions::PyBaseException = self
                         .py()
                         .from_owned_ptr_or_opt($crate::ffi::PyException_GetCause(self.as_ptr()))?;